16/9/24 Update : >Separate method to each class for visibility
